Make `apollo-env` a standard TS package (#1611)
* Make apollo-env a standard TS project

This commit removes all of the extraneous work being done for
compiling apollo-env's global `fetch` types. To accomplish this,
`apollo-env` was converted to a standard TS project using
references, just as the other packages are implemented within this
repo.

* Cleanup extraneous env-ci at the top-level
* Add polyfill setup step for jest
* Remove unused script

I think earlier node versions may need the apollo-env polyfill.
It works on my machine with node v8 but this is a CI test.

* Add apollo-env as a setup file for packages that use the polyfill
* Remove apollo-env from swift codegen dependencies
