+++ package.json
@@ -16,10 +16,9 @@
     "azure": "npm ci && npm run test:ci",
     "clean": "git clean -dfqX -- ./node_modules **/{lib,node_modules}/ ./packages/*/tsconfig.tsbuildinfo ./__tmp__*",
     "postinstall": "npm run build",
-    "build": "lerna run build --stream --scope apollo-env && npm run compile && lerna run build --stream --ignore apollo-env",
-    "compile": "tsc --build tsconfig.json",
-    "compile:clean": "tsc --build tsconfig.json --clean",
-    "watch": "tsc --build tsconfig.json --watch",
+    "build": "tsc --build tsconfig.json",
+    "build:clean": "npm run build -- --clean",
+    "watch": "npm run build -- --watch",
     "test": "jest",
     "test:ci": "npm run test -- --ci --maxWorkers=2",
     "test:ci:debug": "DEBUG=* node --inspect-brk=9001 node_modules/.bin/jest --runInBand",
@@ -49,7 +48,6 @@
     "apollo-env": "file:packages/apollo-env",
     "apollo-graphql": "file:packages/apollo-graphql",
     "apollo-language-server": "file:packages/apollo-language-server",
-    "env-ci": "^4.1.1",
     "vscode-apollo": "file:packages/vscode-apollo"
   },
   "devDependencies": {
