+++ packages/apollo-codegen-flow/package.json
@@ -20,7 +20,6 @@
     "@babel/generator": "7.6.4",
     "@babel/types": "7.6.3",
     "apollo-codegen-core": "file:../apollo-codegen-core",
-    "apollo-env": "file:../apollo-env",
     "change-case": "^3.0.1",
     "common-tags": "^1.5.1",
     "inflected": "^2.0.3"
@@ -28,15 +27,11 @@
   "jest": {
     "preset": "ts-jest",
     "transformIgnorePatterns": [
-      "/node_modules/",
-      "/apollo-env/"
+      "/node_modules/"
     ],
     "testEnvironment": "node",
     "testMatch": [
       "**/__tests__/*.(js|ts)"
-    ],
-    "setupFiles": [
-      "<rootDir>/../apollo-env/lib/index.js"
     ],
     "testPathIgnorePatterns": [
       "<rootDir>/node_modules/",
