+++ packages/apollo-codegen-scala/package.json
@@ -18,7 +18,6 @@
   },
   "dependencies": {
     "apollo-codegen-core": "file:../apollo-codegen-core",
-    "apollo-env": "file:../apollo-env",
     "change-case": "^3.0.1",
     "common-tags": "^1.5.1",
     "inflected": "^2.0.3"
@@ -26,15 +25,11 @@
   "jest": {
     "preset": "ts-jest",
     "transformIgnorePatterns": [
-      "/node_modules/",
-      "/apollo-env/"
+      "/node_modules/"
     ],
     "testEnvironment": "node",
     "testMatch": [
       "**/__tests__/*.(js|ts)"
-    ],
-    "setupFiles": [
-      "<rootDir>/../apollo-env/lib/index.js"
     ],
     "testPathIgnorePatterns": [
       "<rootDir>/node_modules/",
