--- packages/apollo-env/src/fetch/fetch.js
@@ -1,1 +1,0 @@
-export { default as fetch, Request, Response, Headers } from "node-fetch";
