+++ packages/apollo-env/src/index.ts
@@ -1,5 +1,5 @@
 import "./polyfills";
 
 export * from "./typescript-utility-types";
-export * from "../lib/fetch";
+export * from "./fetch";
 export * from "./utils";
