+++ packages/apollo/tsconfig.json
@@ -17,6 +17,7 @@
     { "path": "../apollo-codegen-scala" },
     { "path": "../apollo-codegen-swift" },
     { "path": "../apollo-codegen-typescript" },
+    { "path": "../apollo-env" },
     { "path": "../apollo-graphql" },
     { "path": "../apollo-language-server" },
     { "path": "../apollo-tools" }
