+++ packages/vscode-apollo/tsconfig.json
@@ -6,8 +6,7 @@
   },
   "exclude": ["**/__tests__/*", "**/__mocks__/*", "./lib"],
   "references": [
-    {
-      "path": "../apollo-language-server"
-    }
+    { "path": "../apollo-env" },
+    { "path": "../apollo-language-server" }
   ]
 }
