+++ tsconfig.json
@@ -11,9 +11,10 @@
     { "path": "./packages/apollo-codegen-scala" },
     { "path": "./packages/apollo-codegen-swift" },
     { "path": "./packages/apollo-codegen-typescript" },
+    { "path": "./packages/apollo-env" },
     { "path": "./packages/apollo-graphql" },
     { "path": "./packages/apollo-language-server" },
-    { "path": "./packages/vscode-apollo" },
-    { "path": "./packages/apollo-tools" }
+    { "path": "./packages/apollo-tools" },
+    { "path": "./packages/vscode-apollo" }
   ]
 }
